requirements.txt
streamlit>=1.28.0
numpy>=1.22.0
Pillow>=9.0.0
opencv-python>=4.5.0
deepface>=0.0.79
tf-keras==2.19.0
scipy>=1.8.0
pycryptodome>=3.15.0
fuzzy-extractor==0.1.0